moved Analyzer
and abstraction compatibility fix
