+++ src/Packfire/PDC/Analyzer/Analyzer.php
@@ -9,10 +9,11 @@
  * All rights reserved.
  */
 
-namespace Packfire\PDC;
+namespace Packfire\PDC\Analyzer;
 
 use Packfire\PDC\Report\ReportType;
 use Packfire\PDC\Report\IReport;
+use Packfire\PDC\Toolbelt;
 
 /**
  * Analyzes source code for namespace, class declaration and usage
@@ -20,7 +21,7 @@ use Packfire\PDC\Report\IReport;
  * @author Sam-Mauris Yong <sam@mauris.sg>
  * @copyright 2012 Sam-Mauris Yong Shan Xian <sam@mauris.sg>
  * @license http://www.opensource.org/licenses/BSD-3-Clause The BSD 3-Clause License
- * @package Packfire\PDC
+ * @package Packfire\PDC\Analyzer
  * @since 1.0.4
  * @link https://github.com/packfire/pdc/
  */
